+++ packages/tools/dom/src/attrs.js
@@ -4,7 +4,7 @@
 
 import * as _ from "lamb";
 
-import {joinWithColon, joinWithSemicolon} from "@svizzle/utils";
+import {joinWithColon, joinWithSemicolon, prepend} from "@svizzle/utils";
 
 /**
  * Return a style string from an object
@@ -14,8 +14,8 @@ import {joinWithColon, joinWithSemicolon} from "@svizzle/utils";
  * @return {string} styleString
  *
  * @example
-makeStyle({color: "red", "font-size": "10px"})
-// "color:red;font-size:10px"
+> makeStyle({color: "red", "font-size": "10px"})
+"color:red;font-size:10px"
  *
  * @version 0.1.0
  */
@@ -27,6 +27,26 @@ export const makeStyle = _.pipe([
 ]);
 
 /**
+ * Return a style string with hyphenate CSS variables derived from the keys of the expected object
+ *
+ * @function
+ * @arg {object} object
+ * @return {string} styleString
+ *
+ * @example
+> makeStyleVars({foo: "red", "bar": "10px"})
+"--foo:red;--bar:10px"
+ *
+ * @version 0.4.0
+ */
+export const makeStyleVars = _.pipe([
+	_.skipIf(_.isNil),
+	_.pairs,
+	_.mapWith(_.pipe([joinWithColon, prepend('--')])),
+	joinWithSemicolon
+]);
+
+/**
  * Return a px representation of the received number.
  * Throws an error if the input is not a number.
  *
@@ -34,7 +54,9 @@ export const makeStyle = _.pipe([
  * @arg {number} number
  * @return {string}
  *
- * @example toPx(10) // "10px"
+ * @example
+> toPx(10)
+"10px"
  *
  * @version 0.1.0
  */